CLDR-15990 cleanup xsd and dtd2md code
- separated the markdown and xsd generation
- cleaned up paths and output
